run ictbroadcast campaign with in vtiger crms

+++ modules/MailManager/views/Mail.php
@@ -2,78 +2,93 @@
 /*+**********************************************************************************
  * The contents of this file are subject to the vtiger CRM Public License Version 1.1
  * ("License"); You may not use this file except in compliance with the License
- * The Original Code is:  vtiger CRM Open Source
+ * The Original Code is: vtiger CRM Open Source
  * The Initial Developer of the Original Code is vtiger.
  * Portions created by vtiger are Copyright (C) vtiger.
  * All Rights Reserved.
  ************************************************************************************/
+
 include_once 'include/Webservices/Create.php';
 include_once 'vtlib/Vtiger/Mailer.php';
 include_once 'vtlib/Vtiger/Version.php';
 
-/**
- * Class which controls Mail operation
- */
 class MailManager_Mail_View extends MailManager_Abstract_View {
 
-    /**
-     * Function which processes request for Mail Operations
-     * @global Integer $list_max_entries_per_page - Number of entries per page
-     * @global PearDataBase Instance $adb
-     * @global Users Instance $current_user
-     * @global String $root_directory
-     * @param Vtiger_Request $request
-     * @return MailManager_Response
-     */
-	function process(Vtiger_Request $request) {
-		global $list_max_entries_per_page, $adb, $current_user;
+	public function process(Vtiger_Request $request) {
+		$db = PearDatabase::getInstance();
+		$currentUserModel = Users_Record_Model::getCurrentUserModel();
 		$moduleName = $request->getModule();
-		$response = new Vtiger_Response();
+		$response = new MailManager_Response();
 
 		if ('open' == $this->getOperationArg($request)) {
 			$foldername = $request->get('_folder');
 			$connector = $this->getConnector($foldername);
-			$folder = $connector->folderInstance($foldername);
 
+			$folder = $connector->folderInstance($foldername);
 			$connector->markMailRead($request->get('_msgno'));
 
-			$mail = $connector->openMail($request->get('_msgno'));
-
-			// Get updated count after opening the email
+			$mail = $connector->openMail($request->get('_msgno'), $foldername);
 			$connector->updateFolder($folder, SA_MESSAGES|SA_UNSEEN);
 
 			$viewer = $this->getViewer($request);
 			$viewer->assign('FOLDER', $folder);
 			$viewer->assign('MAIL', $mail);
+			$viewer->assign('USERNAME', $this->mMailboxModel->mUsername);
+			$viewer->assign('ATTACHMENTS', $mail->attachments(false));
+			$body = $mail->body();
+			$inlineAttachments = $mail->inlineAttachments();
+			if(is_array($inlineAttachments)) {
+				foreach($inlineAttachments as $index => $att) {
+					$cid = $att['cid'];
+					$attch_name = Vtiger_MailRecord::__mime_decode($att['filename']);
+					$id = $mail->muid();
+					$src = "index.php?module=MailManager&view=Index&_operation=mail&_operationarg=attachment_dld&_muid=$id&_atname=".urlencode($attch_name);
+					$body = preg_replace('/cid:'.$cid.'/', $src, $body);
+					$inline_cid[$attch_name] = $cid;
+				}
+			}
+			$viewer->assign('INLINE_ATT', $inline_cid);
+			$viewer->assign('BODY', $body);
 			$viewer->assign('MODULE', $moduleName);
 			$uicontent = $viewer->view('MailOpen.tpl', 'MailManager', true);
 
 			$metainfo  = array(
-				'from' => $mail->from(), 'subject' => $mail->subject(),
-				'msgno' => $mail->msgNo(), 'msguid' => $mail->uniqueid(),
-				'folder' => $foldername );
+					'from' => $mail->from(), 'subject' => $mail->subject(),
+					'msgno' => $mail->msgNo(), 'msguid' => $mail->uniqueid(),
+					'folder' => $foldername, 'to' => $mail->to() );
 
 			$response->isJson(true);
 			$response->setResult( array(
-				'folder' => $foldername, 'unread' => $folder->unreadCount(),
-				'ui' => $uicontent, 'meta' => $metainfo )
+					'folder' => $foldername, 'unread' => $folder->unreadCount(),
+					'ui' => $uicontent, 'meta' => $metainfo )
 			);
 
-		} else if ('mark' == $this->getOperationArg($request)) {
+		} else if ('mark' == $this->getOperationArg($request) && $request->validateWriteAccess()) {
+
 			$foldername = $request->get('_folder');
 			$connector = $this->getConnector($foldername);
+
 			$folder = $connector->folderInstance($foldername);
 			$connector->updateFolder($folder, SA_UNSEEN);
 
 			if ('unread' == $request->get('_markas')) {
-				$connector->markMailUnread($request->get('_msgno'));
+				$msgNos = explode(',', $request->get('_msgno'));
+				foreach($msgNos as $msgNo) {
+					$connector->markMailUnread($msgNo);
+				}
+			} else if('read' == $request->get('_markas')) {
+				$msgNos = explode(',', $request->get('_msgno'));
+				foreach($msgNos as $msgNo) {
+					$connector->markMailRead($msgNo);
+				}
 			}
 
 			$response->isJson(true);
 			$response->setResult ( array('folder' => $foldername, 'unread' => $folder->unreadCount()+1,
-                'status' => true, 'msgno' => $request->get('_msgno') ));
+					'status' => true, 'msgno' => $request->get('_msgno') ));
 
-		}else if('delete' == $this->getOperationArg($request)){
+		} else if('delete' == $this->getOperationArg($request)&& $request->validateWriteAccess()) {
+
 			$msg_no = $request->get('_msgno');
 			$foldername = $request->get('_folder');
 			$connector = $this->getConnector($foldername);
@@ -81,7 +96,9 @@ class MailManager_Mail_View extends MailManager_Abstract_View {
 
 			$response->isJson(true);
 			$response->setResult(array('folder' => $foldername,'status'=>true));
-		} else if('move' == $this->getOperationArg($request)){
+
+		} else if('move' == $this->getOperationArg($request) && $request->validateWriteAccess()) {
+
 			$msg_no = $request->get('_msgno');
 			$foldername = $request->get('_folder');
 
@@ -91,11 +108,11 @@ class MailManager_Mail_View extends MailManager_Abstract_View {
 
 			$response->isJson(true);
 			$response->setResult(array('folder' => $foldername,'status'=>true));
-		} else if ('send' == $this->getOperationArg($request)) {
-			require_once 'modules/MailManager/Config.php';
 
-			// This is to handle larger uploads
-			$memory_limit = MailManager_Config::get('MEMORY_LIMIT');
+		} else if ('send' == $this->getOperationArg($request) && $request->validateWriteAccess()) {
+
+			require_once 'modules/MailManager/Config.php';
+			$memory_limit = MailManager_Config_Model::get('MEMORY_LIMIT');
 			ini_set('memory_limit', $memory_limit);
 
 			$to_string = rtrim($request->get('to'), ',');
@@ -104,28 +121,28 @@ class MailManager_Mail_View extends MailManager_Abstract_View {
 			if (!empty($to_string)) {
 				$toArray = explode(',', $to_string);
 				foreach($toArray as $to) {
-					$relatedtos = MailManager::lookupMailInVtiger($to, $current_user);
+					$relatedtos = MailManager::lookupMailInVtiger($to, $currentUserModel);
 					$referenceArray = Array('Contacts','Accounts','Leads');
-					for($j=0;$j<count($referenceArray);$j++){
-						$val=$referenceArray[$j];
+					for($j=0; $j<count($referenceArray); $j++) {
+						$val = $referenceArray[$j];
 						if (!empty($relatedtos) && is_array($relatedtos)) {
 							for($i=0; $i<count($relatedtos); $i++) {
 								if($i == count($relatedtos)-1) {
 									$relateto = vtws_getIdComponents($relatedtos[$i]['record']);
 									$parentIds = $relateto[1]."@1";
-								}elseif($relatedtos[$i]['module'] == $val){
+								} elseif($relatedtos[$i]['module'] == $val) {
 									$relateto = vtws_getIdComponents($relatedtos[$i]['record']);
 									$parentIds = $relateto[1]."@1";
 									break;
 								}
 							}
 						}
-						if(isset ($parentIds)){
+						if(isset ($parentIds)) {
 							break;
 						}
 					}
-					if($parentIds == ''){
-						if(count($relatedtos) > 0){
+					if($parentIds == '') {
+						if(count($relatedtos) > 0) {
 							$relateto = vtws_getIdComponents($relatedtos[0]['record']);
 							$parentIds = $relateto[1]."@1";
 							break;
@@ -137,7 +154,6 @@ class MailManager_Mail_View extends MailManager_Abstract_View {
 					$subject   = $request->get('subject');
 					$body      = $request->get('body');
 
-					//Restrict this for users module
 					if($relateto[1]!= NULL && $relateto[0] != '19') {
 						$entityId = $relateto[1];
 						$parent_module = getSalesEntityType($entityId);
@@ -148,17 +164,17 @@ class MailManager_Mail_View extends MailManager_Abstract_View {
 					}
 
 					$fromEmail = $connector->getFromEmailAddress();
-					$userFullName = getFullNameFromArray('Users', $current_user->column_fields);
-					$userId = $current_user->id;
+					$userFullName = getFullNameFromArray('Users', $currentUserModel->getData());
+					$userId = $currentUserModel->getId();
 
 					$mailer = new Vtiger_Mailer();
 					$mailer->IsHTML(true);
-					$mailer->ConfigSenderInfo($fromEmail, $userFullName, $current_user->email1);
+					$mailer->ConfigSenderInfo($fromEmail, $userFullName, $currentUserModel->get('email1'));
 					$mailer->Subject = $subject;
 					$mailer->Body = $description;
 					$mailer->addSignature($userId);
-		            if($mailer->Signature != '') {
-		               $mailer->Body.= $mailer->Signature;
+					if($mailer->Signature != '') {
+						$mailer->Body.= $mailer->Signature;
 					}
 
 					$ccs = empty($cc_string)? array() : explode(',', $cc_string);
@@ -166,15 +182,13 @@ class MailManager_Mail_View extends MailManager_Abstract_View {
 					$emailId = $request->get('emailid');
 
 					$attachments = $connector->getAttachmentDetails($emailId);
-
 					$mailer->AddAddress($to);
 					foreach($ccs as $cc) $mailer->AddCC($cc);
 					foreach($bccs as $bcc)$mailer->AddBCC($bcc);
-					global $root_directory;
 
 					if(is_array($attachments)) {
-						foreach($attachments as $attachment){
-							$fileNameWithPath = $root_directory.$attachment['path'].$attachment['fileid']."_".$attachment['attachment'];
+						foreach($attachments as $attachment) {
+							$fileNameWithPath = vglobal('root_directory').$attachment['path'].$attachment['fileid']."_".$attachment['attachment'];
 							if(is_file($fileNameWithPath)) {
 								$mailer->AddAttachment($fileNameWithPath, $attachment['attachment']);
 							}
@@ -184,7 +198,7 @@ class MailManager_Mail_View extends MailManager_Abstract_View {
 
 					if ($status === true) {
 						$email = CRMEntity::getInstance('Emails');
-						$email->column_fields['assigned_user_id'] = $current_user->id;
+						$email->column_fields['assigned_user_id'] = $currentUserModel->getId();
 						$email->column_fields['date_start'] = date('Y-m-d');
 						$email->column_fields['time_start'] = date('H:i');
 						$email->column_fields['parent_id'] = $parentIds;
@@ -196,6 +210,7 @@ class MailManager_Mail_View extends MailManager_Abstract_View {
 						$email->column_fields['ccmail'] = $cc_string;
 						$email->column_fields['bccmail'] = $bcc_string;
 						$email->column_fields['email_flag'] = 'SENT';
+
 						if(empty($emailId)) {
 							$email->save('Emails');
 						} else {
@@ -204,17 +219,16 @@ class MailManager_Mail_View extends MailManager_Abstract_View {
 							$email->save('Emails');
 						}
 
-						// This is added since the Emails save_module is not handling this
-						global $adb;
 						$realid = explode("@", $parentIds);
 						$mycrmid = $realid[0];
 						$params = array($mycrmid, $email->id);
+
 						if ($realid[1] == -1) {
-							$adb->pquery('DELETE FROM vtiger_salesmanactivityrel WHERE smid=? AND activityid=?',$params);
-							$adb->pquery('INSERT INTO vtiger_salesmanactivityrel VALUES (?,?)', $params);
+							$db->pquery('DELETE FROM vtiger_salesmanactivityrel WHERE smid=? AND activityid=?',$params);
+							$db->pquery('INSERT INTO vtiger_salesmanactivityrel VALUES (?,?)', $params);
 						} else {
-							$adb->pquery('DELETE FROM vtiger_seactivityrel WHERE crmid=? AND activityid=?', $params);
-							$adb->pquery('INSERT INTO vtiger_seactivityrel VALUES (?,?)', $params);
+							$db->pquery('DELETE FROM vtiger_seactivityrel WHERE crmid=? AND activityid=?', $params);
+							$db->pquery('INSERT INTO vtiger_seactivityrel VALUES (?,?)', $params);
 						}
 					}
 				}
@@ -227,42 +241,45 @@ class MailManager_Mail_View extends MailManager_Abstract_View {
 				$response->isJson(true);
 				$response->setError(112, 'please verify outgoing server.');
 			}
+
 		} else if ('attachment_dld' == $this->getOperationArg($request)) {
-			$attachmentName = $request->get('_atname');
+			$attachmentName = $request->getRaw('_atname');
 			$attachmentName= str_replace(' ', '_', $attachmentName);
 
-			if (MailManager_Utils::allowedFileExtension($attachmentName)) {
+			if (MailManager_Utils_Helper::allowedFileExtension($attachmentName)) {
 				// This is to handle larger uploads
-				$memory_limit = MailManager_Config::get('MEMORY_LIMIT');
+				$memory_limit = MailManager_Config_Model::get('MEMORY_LIMIT');
 				ini_set('memory_limit', $memory_limit);
 
 				$mail = new MailManager_Message_Model(false, false);
 				$mail->readFromDB($request->get('_muid'));
 				$attachment = $mail->attachments(true, $attachmentName);
-
-                if($attachment[$attachmentName]) {
-                    // Send as downloadable
-                    header("Content-type: application/octet-stream");
-                    header("Pragma: public");
-                    header("Cache-Control: private");
-                    header("Content-Disposition: attachment; filename=$attachmentName");
-                    echo $attachment[$attachmentName];
-                } else {
-                    header("Content-Disposition: attachment; filename=INVALIDFILE");
-                    echo "";
-                }
+				//As we are sending attachment name, it will return only that attachment details
+				if($attachment[0]['data']) {
+					header("Content-type: application/octet-stream");
+					header("Pragma: public");
+					header("Cache-Control: private");
+					header("Content-Disposition: attachment; filename=\"$attachmentName\"");
+					echo $attachment[0]['data'];
+				} else {
+					header("Content-Disposition: attachment; filename=INVALIDFILE");
+					echo "";
+				}
 			} else {
 				header("Content-Disposition: attachment; filename=INVALIDFILE");
 				echo "";
 			}
 			flush();
 			exit;
-		} elseif('getdraftmail' == $this->getOperationArg($request)) {
+		} elseif('getdraftmail' == $this->getOperationArg($request) && $request->validateWriteAccess()) {
+
 			$connector = $this->getConnector('__vt_drafts');
 			$draftMail = $connector->getDraftMail($request);
 			$response->isJson(true);
 			$response->setResult(array($draftMail));
-		} elseif('save' == $this->getOperationArg($request)){
+
+		} elseif('save' == $this->getOperationArg($request) && $request->validateWriteAccess()) {
+
 			$connector = $this->getConnector('__vt_drafts');
 			$draftId = $connector->saveDraft($request);
 
@@ -272,28 +289,45 @@ class MailManager_Mail_View extends MailManager_Abstract_View {
 			} else {
 				$response->setResult( array('success'=> false,'error'=>"Draft was not saved") );
 			}
-		} elseif('deleteAttachment' == $this->getOperationArg($request)) {
+
+		} elseif('deleteAttachment' == $this->getOperationArg($request) && $request->validateWriteAccess()) {
+
 			$connector = $this->getConnector('__vt_drafts');
 			$deleteResponse = $connector->deleteAttachment($request);
 
 			$response->isJson(true);
 			$response->setResult(array('success'=> $deleteResponse));
 
-		} elseif('forward' == $this->getOperationArg($request)) {
+		} elseif('forward' == $this->getOperationArg($request) && $request->validateWriteAccess()) {
+
 			$messageId = $request->get('messageid');
 			$folderName = $request->get('folder');
 
 			$connector = $this->getConnector($folderName);
-			$mail = $connector->openMail($messageId);
-
+			$mail = $connector->openMail($messageId, $folderName);
 			$attachments = $mail->attachments(true);
 
+			$inlineAttachments = $mail->inlineAttachments();
 			$draftConnector = $this->getConnector('__vt_drafts');
 			$draftId = $draftConnector->saveDraft($request);
 
+			if(is_array($inlineAttachments)) {
+				foreach($inlineAttachments as $index => $att) {
+					$cid = $att['cid'];
+					$attch_name = imap_utf8($att['filename']);
+					$id = $mail->muid();
+					$src = "index.php?module=MailManager&view=Index&_operation=mail&_operationarg=attachment_dld&_muid=$id&_atname=$attch_name";
+					$body = preg_replace('/cid:'.$cid.'/', $src, $body);
+					$inline_cid[$attch_name] = $cid;
+				}
+			}
+
 			if (!empty($attachments)) {
-				foreach($attachments as $aName => $aValue) {
-					$attachInfo = $mail->__SaveAttachmentFile($aName, $aValue);
+				foreach($attachments as $aValue) {
+					if(!empty($inline_cid[$aValue['filename']])) {
+						continue;
+					}
+					$attachInfo = $mail->__SaveAttachmentFile($aValue['filename'], $aValue['data']);
 					if(is_array($attachInfo) && !empty($attachInfo) && $attachInfo['size'] > 0) {
 
 						if(!MailManager::checkModuleWriteAccessForCurrentUser('Documents')) return;
@@ -305,27 +339,26 @@ class MailManager_Mail_View extends MailManager_Abstract_View {
 						$document->column_fields['filelocationtype'] = 'I';
 						$document->column_fields['folderid']         = 1; // Default Folder
 						$document->column_fields['filesize']		 = $attachInfo['size'];
-						$document->column_fields['assigned_user_id'] = $current_user->id;
+						$document->column_fields['assigned_user_id'] = $currentUserModel->getId();
 						$document->save('Documents');
 
-						//save doc-attachment relation
 						$draftConnector->saveAttachmentRel($document->id, $attachInfo['attachid']);
-
-						//save email-doc relation
 						$draftConnector->saveEmailDocumentRel($draftId, $document->id);
-
-						//save email-attachment relation
 						$draftConnector->saveAttachmentRel($draftId, $attachInfo['attachid']);
 
 						$attachmentInfo[] = array('name'=>$attachInfo['name'], 'size'=>$attachInfo['size'], 'emailid'=>$draftId, 'docid'=>$document->id);
 					}
-					unset($aValue);
+					unset($aValue['data']);
 				}
 			}
 			$response->isJson(true);
 			$response->setResult(array('attachments'=>$attachmentInfo, 'emailid'=>$draftId));
 		}
 		return $response;
+	}
+
+	public function validateRequest(Vtiger_Request $request) {
+		return $request->validateReadAccess();
 	}
 }
 ?>
